Split events file into more specific files.
