+++ pkg/odo/events/listener.go
@@ -1,0 +1,7 @@
+package events
+
+type Listener interface {
+	OnEvent(event Event) error
+	OnAbort(abortError *EventCausedAbortError)
+	Name() string
+}
